Adicionado restrições em produtos, itens do pedido e pedido

+++ src/main/java/com/senior/application/service/impl/ProdutoServicoServiceImpl.java
@@ -4,13 +4,18 @@ import java.util.List;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.senior.application.constants.MensagemServidor;
+import com.senior.application.exceptions.http.InternalErrorException;
+import com.senior.application.model.ItemPedido;
 import com.senior.application.model.ProdutoServico;
+import com.senior.application.repository.ItemPedidoRepository;
 import com.senior.application.repository.ProdutoServicoRepository;
 import com.senior.application.service.ProdutoServicoService;
 import com.senior.application.util.OptionalUtil;
@@ -21,6 +26,9 @@ public class ProdutoServicoServiceImpl implements ProdutoServicoService {
 
 	@Autowired
 	private ProdutoServicoRepository produtoServicoRepository;
+
+	@Autowired
+	private ItemPedidoRepository itemPedidoRepository;
 
 	@Override
 	@Transactional
@@ -44,6 +52,7 @@ public class ProdutoServicoServiceImpl implements ProdutoServicoService {
 
 		produtoServico.setDescricao(cadastroProdutoServicoDTO.getDescricao());
 		produtoServico.setTipo(cadastroProdutoServicoDTO.getTipo());
+		produtoServico.setSituacao(cadastroProdutoServicoDTO.getSituacao());
 
 		return produtoServicoRepository.save(produtoServico);
 	}
@@ -52,6 +61,14 @@ public class ProdutoServicoServiceImpl implements ProdutoServicoService {
 	public void deleteProdutoServico(UUID id) {
 		// Lança NotFoundException caso o id não exista
 		readProdutoServico(id);
+		
+		ItemPedido itemPedido = new ItemPedido();
+		itemPedido.getItemPedidoId().setIdProdutoServico(id);
+		Example<ItemPedido> example = Example.of(itemPedido, ExampleMatcher.matching());
+		
+		if (itemPedidoRepository.count(example) > 0) {
+			throw new InternalErrorException(MensagemServidor.EXCLUIR_PRODUTO_SERVICO_VINCULADO_PEDIDO);
+		}
 		produtoServicoRepository.deleteById(id);
 	}
